gwproxy: Move the main epoll event loop into epoll.c
A preparation to migrate epoll specific code to epoll.c. Currently, the
`handle_event()` function is kept in gwproxy.c and is exported as
`gwp_ctx_handle_event_epoll()` because it's too big to be migrated at
once to epoll.c.
