+++ src/gwproxy/ev/epoll.h
@@ -9,5 +9,6 @@
 
 int gwp_ctx_init_thread_epoll(struct gwp_wrk *w);
 void gwp_ctx_free_thread_epoll(struct gwp_wrk *w);
+int gwp_ctx_thread_entry_epoll(struct gwp_wrk *w);
 
 #endif /* #ifndef GWPROXY__EV__EPOLL_H */
